requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
faker
matplotlib
seaborn
pyspark
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
requests